Add interfaces for services, improve README, remove unused code, small refactors
