+++ api/src/main/java/io/github/kkw/api/controllers/LoginController.java
@@ -9,9 +9,7 @@ import io.github.kkw.api.model.ClientId;
 import io.github.kkw.api.model.Profile;
 import io.github.kkw.api.services.LoginService;
 import org.springframework.http.HttpStatus;
-import org.springframework.lang.NonNull;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
