--- api/src/main/java/io/github/kkw/api/exceptions/AvailableCinemaHallsNotFoundException.java
@@ -1,5 +1,0 @@
-package io.github.kkw.api.exceptions;
-
-public class AvailableCinemaHallsNotFoundException extends Exception{
-    public AvailableCinemaHallsNotFoundException(final String message){super(message);}
-}
