--- api/src/main/java/io/github/kkw/api/exceptions/SpecialCodeAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package io.github.kkw.api.exceptions;
-
-public class SpecialCodeAlreadyExistsException extends Exception{
-    public SpecialCodeAlreadyExistsException(String message) {
-        super(message);
-    }
-}
