--- api/src/main/java/io/github/kkw/api/exceptions/SpecialOffersNotFoundException.java
@@ -1,7 +1,0 @@
-package io.github.kkw.api.exceptions;
-
-public class SpecialOffersNotFoundException extends Exception{
-    public SpecialOffersNotFoundException(String message) {
-        super(message);
-    }
-}
