+++ api/src/main/java/io/github/kkw/api/services/HallReservationService.java
@@ -1,10 +1,5 @@
 package io.github.kkw.api.services;
 
-import io.github.kkw.api.db.HallReservationRepository;
-import io.github.kkw.api.db.MovieRepository;
-import io.github.kkw.api.db.dto.HallEntity;
-import io.github.kkw.api.db.dto.HallReservationEntity;
-import io.github.kkw.api.exceptions.AvailableCinemaHallsNotFoundException;
 import io.github.kkw.api.exceptions.CinemaHallsNotFoundException;
 import io.github.kkw.api.exceptions.DateTooSoonException;
 import io.github.kkw.api.exceptions.HallNotFoundException;
@@ -14,65 +9,46 @@ import io.github.kkw.api.model.ClientId;
 import io.github.kkw.api.model.Hall;
 import io.github.kkw.api.model.HallReservation;
 import io.github.kkw.api.model.ReservationId;
-import org.springframework.stereotype.Service;
 
 import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-public class HallReservationService {
-    private final HallReservationRepository hallReservationRepository;
+public interface HallReservationService {
+    /**
+     *
+     * @param clientId ID of the calling client
+     * @param hallId ID of the hall to reserve
+     * @param date date to reserve the hall on
+     * @return reservation ID of newly created hall reservation
+     * @throws HallReservedException when hall is already reserved
+     * @throws HallNotFoundException when hall is not found
+     * @throws DateTooSoonException reservations for halls can only be done 2 weeks in advance
+     */
+    ReservationId createHallReservation(ClientId clientId, int hallId, Instant date) throws HallReservedException, HallNotFoundException, DateTooSoonException;
 
-    public HallReservationService(HallReservationRepository hallReservationRepository) {
-        this.hallReservationRepository = hallReservationRepository;
-    }
+    /**
+     *
+     * @param clientId ID of the calling client
+     * @param reservationId ID of the reservation to delete
+     * @throws ReservationNotFoundException when reservation with this ID was not found
+     */
+    void deleteHallReservation(ClientId clientId, ReservationId reservationId) throws ReservationNotFoundException;
 
-    public ReservationId createHallReservation(ClientId clientId, int hallId, Instant date) throws HallReservedException, HallNotFoundException, DateTooSoonException {
-        long daysFromNowToDate = Instant.now().until(date, ChronoUnit.DAYS);
-        if (daysFromNowToDate < 14) {
-            throw new DateTooSoonException("Date must be at least 14 days in the future");
-        }
-        if (hallReservationRepository.isReserved(hallId, date)) {
-            throw new HallReservedException("Hall with this ID is already reserved for chosen day");
-        }
-        if (!hallReservationRepository.isHallExists(hallId)) {
-            throw new HallNotFoundException("Hall with this ID not found");
-        }
-        hallReservationRepository.createHallReservation(clientId.getClientId(), hallId, date);
-        return new ReservationId(hallReservationRepository.getHallReservationId(clientId.getClientId(), hallId, date));
-    }
+    /**
+     *
+     * @param date date to show available halls for
+     * @return list of all halls that are available to reserve on chosen date
+     * @throws CinemaHallsNotFoundException when no available cinema halls were found
+     * @throws DateTooSoonException reservations for halls can only be done 2 weeks in advance
+     */
+    List<Hall> showAvailableCinemaHalls(final Instant date) throws CinemaHallsNotFoundException, DateTooSoonException;
 
-    public void deleteHallReservation(ClientId clientId, ReservationId reservationId) throws ReservationNotFoundException {
-        if (hallReservationRepository.reservationDoesntExist(clientId.getClientId(), reservationId.getReservationId())) {
-            throw new ReservationNotFoundException("Reservation with this ID not found");
-        }
-        hallReservationRepository.deleteHallReservation(clientId.getClientId(), reservationId.getReservationId());
-    }
-
-    public List<Hall> showCinemaHalls() throws CinemaHallsNotFoundException {
-        List<HallEntity> hallEntities = hallReservationRepository.showCinemaHalls();
-        if(hallEntities.isEmpty()){
-            throw new CinemaHallsNotFoundException("Cannot find any cinema halls");
-        }
-        return hallEntities.stream().map(Hall::new).collect(Collectors.toList());
-    }
-
-    public List<Hall> showAvailableCinemaHalls(final Instant date) throws AvailableCinemaHallsNotFoundException, DateTooSoonException {
-        long daysFromNowToDate = Instant.now().until(date, ChronoUnit.DAYS);
-        if (daysFromNowToDate < 14) {
-            throw new DateTooSoonException("Date must be at least 14 days in the future");
-        }
-        List<HallEntity> hallEntities = hallReservationRepository.showAvailableCinemaHalls(date);
-        if(hallEntities.isEmpty()){
-            throw new AvailableCinemaHallsNotFoundException("Cannot find any cinema halls in this term");
-        }
-        return hallEntities.stream().map(Hall::new).collect(Collectors.toList());
-    }
-
-    public List<HallReservation> showReservations(ClientId clientId, Instant from, Instant to) {
-        List<HallReservationEntity> allReservations = hallReservationRepository.getReservations(clientId.getClientId(), from, to);
-        return allReservations.stream().map(HallReservation::new).collect(Collectors.toList());
-    }
+    /**
+     *
+     * @param clientId ID of the calling client
+     * @param from date to show reservations from
+     * @param to date to show reservation to
+     * @return list of all hall reservations for chosen client in chosen time frame
+     */
+    List<HallReservation> showReservations(ClientId clientId, Instant from, Instant to);
 }
