+++ docs/architecture.md
@@ -1,0 +1,2 @@
+## Database schema
+![Database schema](files/database_schema.jpg)
