+++ front/src/components/ApiUtils.js
@@ -1,21 +1,26 @@
 export const baseUrl = "http://localhost:8888"
-export const loginSubUrl = "/login"
+// login
 export const registerSubUrl = "/register"
+export const loginSubUrl = "/login"
 export const profileSubUrl = "/showProfile"
+// admin
+export const verifyAdminSubUrl = "/verifyAdmin"
+export const statisticsGeneralSubUrl = "/showStatistics"
+export const statisticsMovieSubUrl = "/showStatistics/movie"
+export const statisticsHallSubUrl = "/showStatistics/hall"
+// seat reservations
 export const moviesSubUrl = "/showMovies"
 export const availableSeatsSubUrl = "/showFreeSeats"
 export const reserveSeatSubUrl = "/reserveSeat"
 export const showSeatReservationsSubUrl = "/showReservations/seat"
+// hall reservations
+export const availableCinemaHallsUrl = "/showAvailableCinemaHalls"
+export const reserveHallSubUrl = "/reserveHall"
 export const showHallReservationsSubUrl = "/showReservations/hall"
+// payments
 export const payForSeatSubUrl = "/payForSeat"
 export const payAdvanceForHallSubUrl = "/payAdvanceForHall"
 export const payForHallSubUrl = "/payForHall"
-export const verifyAdminSubUrl = "/verifyAdmin"
-export const availableCinemaHallsUrl = "/showAvailableCinemaHalls"
-export const reserveHallSubUrl = "/reserveHall"
-export const statisticsGeneralSubUrl = "/showStatistics"
-export const statisticsMovieSubUrl = "/showStatistics/movie"
-export const statisticsHallSubUrl = "/showStatistics/hall"
 
 export function buildUrl(subUrl, pathVars, urlParams) {
     const url = [baseUrl, subUrl];
@@ -43,7 +48,7 @@ export function callCrsApi(url, params, onSuccess, onFail) {
             }
         }))
         .then((data) => {
-            if (data.hasOwnProperty("status") && data.hasOwnProperty("message")) {
+            if (data.hasOwnProperty("httpStatus") && data.hasOwnProperty("message")) {
                 onFail(data)
             } else {
                 onSuccess(data)
