+++ gradle.properties
@@ -1,0 +1,1 @@
+version=0.0.2
